Remove unnecessary functions from headers
I've noticed that actsvg has most of its code in header files. While
this is well managed using the `static` and `inline` keywords, it still
leads to some rather undesirable effects. For example, I ran into an
issue where a detray CUDA file was crashing because of some code in
actsvg, that that kind of very deep dependency is not very maintainable.
For that reason, I suggest we move as much of the code from actsvg into
sources files. This will help reduce compilation times and memory usage.
