+++ core/include/actsvg/core/generators.hpp
@@ -26,16 +26,8 @@ namespace generators {
  *
  * @return a vector of phi values for the arc
  */
-inline std::vector<scalar> phi_values(scalar start_phi, scalar end_phi,
-                                      unsigned int lseg) {
-    std::vector<scalar> values;
-    values.reserve(lseg + 1);
-    scalar step_phi = (end_phi - start_phi) / lseg;
-    for (unsigned int istep = 0; istep <= lseg; ++istep) {
-        values.push_back(start_phi + istep * step_phi);
-    }
-    return values;
-}
+std::vector<scalar> phi_values(scalar start_phi, scalar end_phi,
+                               unsigned int lseg);
 
 /** Generate a contour for a sector
  *
@@ -49,31 +41,9 @@ inline std::vector<scalar> phi_values(scalar start_phi, scalar end_phi,
  * @note start/end are respected in terms of orientation
  *
  **/
-inline std::vector<point2> sector_contour(scalar inner_r, scalar outer_r,
-                                          scalar start_phi, scalar end_phi,
-                                          unsigned int lseg = 16) {
-
-    // Re-bound phi
-    if (start_phi > 0. and end_phi < 0.) {
-        end_phi += 2 * M_PI;
-    }
-
-    auto inner_phi = phi_values(end_phi, start_phi, lseg);
-    auto outer_phi = phi_values(start_phi, end_phi, lseg);
-
-    std::vector<point2> sector_vertices;
-    sector_vertices.reserve(inner_phi.size() + outer_phi.size());
-    for (auto iphi : inner_phi) {
-        sector_vertices.push_back(
-            {inner_r * std::cos(iphi), inner_r * std::sin(iphi)});
-    }
-    for (auto ophi : outer_phi) {
-        sector_vertices.push_back(
-            {outer_r * std::cos(ophi), outer_r * std::sin(ophi)});
-    }
-    return sector_vertices;
-}
-
+std::vector<point2> sector_contour(scalar inner_r, scalar outer_r,
+                                   scalar start_phi, scalar end_phi,
+                                   unsigned int lseg = 16);
 }  // namespace generators
 
 }  // namespace actsvg
