+++ data/include/actsvg/data/odd_pixel_barrel.hpp
@@ -18,919 +18,11 @@ namespace actsvg {
 /// This file is generated from the Open Data detector and
 /// depicts one Pixel barrel layer
 namespace data {
-
-static std::vector<std::array<scalar, 3u>> odd_pixel_barrel = {
-    {69.2185, -8.55507, -504},    {67.2074, 8.12411, -504},
-    {67.2074, 8.12411, -432},     {69.2185, -8.55507, -432},
-    {69.2185, -8.55507, -431.5},  {67.2074, 8.12411, -431.5},
-    {67.2074, 8.12411, -359.5},   {69.2185, -8.55507, -359.5},
-    {69.2185, -8.55507, 221},     {67.2074, 8.12411, 221},
-    {67.2074, 8.12411, 293},      {69.2185, -8.55507, 293},
-    {69.2185, -8.55507, 293.5},   {67.2074, 8.12411, 293.5},
-    {67.2074, 8.12411, 365.5},    {69.2185, -8.55507, 365.5},
-    {69.2185, -8.55507, 366},     {67.2074, 8.12411, 366},
-    {67.2074, 8.12411, 438},      {69.2185, -8.55507, 438},
-    {69.2185, -8.55507, 438.5},   {67.2074, 8.12411, 438.5},
-    {67.2074, 8.12411, 510.5},    {69.2185, -8.55507, 510.5},
-    {69.2185, -8.55507, -359},    {67.2074, 8.12411, -359},
-    {67.2074, 8.12411, -287},     {69.2185, -8.55507, -287},
-    {69.2185, -8.55507, -286.5},  {67.2074, 8.12411, -286.5},
-    {67.2074, 8.12411, -214.5},   {69.2185, -8.55507, -214.5},
-    {69.2185, -8.55507, -214},    {67.2074, 8.12411, -214},
-    {67.2074, 8.12411, -142},     {69.2185, -8.55507, -142},
-    {69.2185, -8.55507, -141.5},  {67.2074, 8.12411, -141.5},
-    {67.2074, 8.12411, -69.5},    {69.2185, -8.55507, -69.5},
-    {69.2185, -8.55507, -69},     {67.2074, 8.12411, -69},
-    {67.2074, 8.12411, 3},        {69.2185, -8.55507, 3},
-    {69.2185, -8.55507, 3.5},     {67.2074, 8.12411, 3.5},
-    {67.2074, 8.12411, 75.5},     {69.2185, -8.55507, 75.5},
-    {69.2185, -8.55507, 76},      {67.2074, 8.12411, 76},
-    {67.2074, 8.12411, 148},      {69.2185, -8.55507, 148},
-    {69.2185, -8.55507, 148.5},   {67.2074, 8.12411, 148.5},
-    {67.2074, 8.12411, 220.5},    {69.2185, -8.55507, 220.5},
-    {69.5575, 5.11317, -504},     {64.3311, 21.0795, -504},
-    {64.3311, 21.0795, -432},     {69.5575, 5.11317, -432},
-    {69.5575, 5.11317, -431.5},   {64.3311, 21.0795, -431.5},
-    {64.3311, 21.0795, -359.5},   {69.5575, 5.11317, -359.5},
-    {69.5575, 5.11317, 221},      {64.3311, 21.0795, 221},
-    {64.3311, 21.0795, 293},      {69.5575, 5.11317, 293},
-    {69.5575, 5.11317, 293.5},    {64.3311, 21.0795, 293.5},
-    {64.3311, 21.0795, 365.5},    {69.5575, 5.11317, 365.5},
-    {69.5575, 5.11317, 366},      {64.3311, 21.0795, 366},
-    {64.3311, 21.0795, 438},      {69.5575, 5.11317, 438},
-    {69.5575, 5.11317, 438.5},    {64.3311, 21.0795, 438.5},
-    {64.3311, 21.0795, 510.5},    {69.5575, 5.11317, 510.5},
-    {69.5575, 5.11317, -359},     {64.3311, 21.0795, -359},
-    {64.3311, 21.0795, -287},     {69.5575, 5.11317, -287},
-    {69.5575, 5.11317, -286.5},   {64.3311, 21.0795, -286.5},
-    {64.3311, 21.0795, -214.5},   {69.5575, 5.11317, -214.5},
-    {69.5575, 5.11317, -214},     {64.3311, 21.0795, -214},
-    {64.3311, 21.0795, -142},     {69.5575, 5.11317, -142},
-    {69.5575, 5.11317, -141.5},   {64.3311, 21.0795, -141.5},
-    {64.3311, 21.0795, -69.5},    {69.5575, 5.11317, -69.5},
-    {69.5575, 5.11317, -69},      {64.3311, 21.0795, -69},
-    {64.3311, 21.0795, 3},        {69.5575, 5.11317, 3},
-    {69.5575, 5.11317, 3.5},      {64.3311, 21.0795, 3.5},
-    {64.3311, 21.0795, 75.5},     {69.5575, 5.11317, 75.5},
-    {69.5575, 5.11317, 76},       {64.3311, 21.0795, 76},
-    {64.3311, 21.0795, 148},      {69.5575, 5.11317, 148},
-    {69.5575, 5.11317, 148.5},    {64.3311, 21.0795, 148.5},
-    {64.3311, 21.0795, 220.5},    {69.5575, 5.11317, 220.5},
-    {-18.5849, 67.2235, -504},    {-33.2248, 58.9825, -504},
-    {-33.2248, 58.9825, -432},    {-18.5849, 67.2235, -432},
-    {-18.5849, 67.2235, -431.5},  {-33.2248, 58.9825, -431.5},
-    {-33.2248, 58.9825, -359.5},  {-18.5849, 67.2235, -359.5},
-    {-18.5849, 67.2235, 221},     {-33.2248, 58.9825, 221},
-    {-33.2248, 58.9825, 293},     {-18.5849, 67.2235, 293},
-    {-18.5849, 67.2235, 293.5},   {-33.2248, 58.9825, 293.5},
-    {-33.2248, 58.9825, 365.5},   {-18.5849, 67.2235, 365.5},
-    {-18.5849, 67.2235, 366},     {-33.2248, 58.9825, 366},
-    {-33.2248, 58.9825, 438},     {-18.5849, 67.2235, 438},
-    {-18.5849, 67.2235, 438.5},   {-33.2248, 58.9825, 438.5},
-    {-33.2248, 58.9825, 510.5},   {-18.5849, 67.2235, 510.5},
-    {-18.5849, 67.2235, -359},    {-33.2248, 58.9825, -359},
-    {-33.2248, 58.9825, -287},    {-18.5849, 67.2235, -287},
-    {-18.5849, 67.2235, -286.5},  {-33.2248, 58.9825, -286.5},
-    {-33.2248, 58.9825, -214.5},  {-18.5849, 67.2235, -214.5},
-    {-18.5849, 67.2235, -214},    {-33.2248, 58.9825, -214},
-    {-33.2248, 58.9825, -142},    {-18.5849, 67.2235, -142},
-    {-18.5849, 67.2235, -141.5},  {-33.2248, 58.9825, -141.5},
-    {-33.2248, 58.9825, -69.5},   {-18.5849, 67.2235, -69.5},
-    {-18.5849, 67.2235, -69},     {-33.2248, 58.9825, -69},
-    {-33.2248, 58.9825, 3},       {-18.5849, 67.2235, 3},
-    {-18.5849, 67.2235, 3.5},     {-33.2248, 58.9825, 3.5},
-    {-33.2248, 58.9825, 75.5},    {-18.5849, 67.2235, 75.5},
-    {-18.5849, 67.2235, 76},      {-33.2248, 58.9825, 76},
-    {-33.2248, 58.9825, 148},     {-18.5849, 67.2235, 148},
-    {-18.5849, 67.2235, 148.5},   {-33.2248, 58.9825, 148.5},
-    {-33.2248, 58.9825, 220.5},   {-18.5849, 67.2235, 220.5},
-    {-31.3425, 62.306, -504},     {-44.0934, 51.3674, -504},
-    {-44.0934, 51.3674, -432},    {-31.3425, 62.306, -432},
-    {-31.3425, 62.306, -431.5},   {-44.0934, 51.3674, -431.5},
-    {-44.0934, 51.3674, -359.5},  {-31.3425, 62.306, -359.5},
-    {-31.3425, 62.306, 221},      {-44.0934, 51.3674, 221},
-    {-44.0934, 51.3674, 293},     {-31.3425, 62.306, 293},
-    {-31.3425, 62.306, 293.5},    {-44.0934, 51.3674, 293.5},
-    {-44.0934, 51.3674, 365.5},   {-31.3425, 62.306, 365.5},
-    {-31.3425, 62.306, 366},      {-44.0934, 51.3674, 366},
-    {-44.0934, 51.3674, 438},     {-31.3425, 62.306, 438},
-    {-31.3425, 62.306, 438.5},    {-44.0934, 51.3674, 438.5},
-    {-44.0934, 51.3674, 510.5},   {-31.3425, 62.306, 510.5},
-    {-31.3425, 62.306, -359},     {-44.0934, 51.3674, -359},
-    {-44.0934, 51.3674, -287},    {-31.3425, 62.306, -287},
-    {-31.3425, 62.306, -286.5},   {-44.0934, 51.3674, -286.5},
-    {-44.0934, 51.3674, -214.5},  {-31.3425, 62.306, -214.5},
-    {-31.3425, 62.306, -214},     {-44.0934, 51.3674, -214},
-    {-44.0934, 51.3674, -142},    {-31.3425, 62.306, -142},
-    {-31.3425, 62.306, -141.5},   {-44.0934, 51.3674, -141.5},
-    {-44.0934, 51.3674, -69.5},   {-31.3425, 62.306, -69.5},
-    {-31.3425, 62.306, -69},      {-44.0934, 51.3674, -69},
-    {-44.0934, 51.3674, 3},       {-31.3425, 62.306, 3},
-    {-31.3425, 62.306, 3.5},      {-44.0934, 51.3674, 3.5},
-    {-44.0934, 51.3674, 75.5},    {-31.3425, 62.306, 75.5},
-    {-31.3425, 62.306, 76},       {-44.0934, 51.3674, 76},
-    {-44.0934, 51.3674, 148},     {-31.3425, 62.306, 148},
-    {-31.3425, 62.306, 148.5},    {-44.0934, 51.3674, 148.5},
-    {-44.0934, 51.3674, 220.5},   {-31.3425, 62.306, 220.5},
-    {-42.8955, 54.9942, -504},    {-53.2674, 41.7782, -504},
-    {-53.2674, 41.7782, -432},    {-42.8955, 54.9942, -432},
-    {-42.8955, 54.9942, -431.5},  {-53.2674, 41.7782, -431.5},
-    {-53.2674, 41.7782, -359.5},  {-42.8955, 54.9942, -359.5},
-    {-42.8955, 54.9942, 221},     {-53.2674, 41.7782, 221},
-    {-53.2674, 41.7782, 293},     {-42.8955, 54.9942, 293},
-    {-42.8955, 54.9942, 293.5},   {-53.2674, 41.7782, 293.5},
-    {-53.2674, 41.7782, 365.5},   {-42.8955, 54.9942, 365.5},
-    {-42.8955, 54.9942, 366},     {-53.2674, 41.7782, 366},
-    {-53.2674, 41.7782, 438},     {-42.8955, 54.9942, 438},
-    {-42.8955, 54.9942, 438.5},   {-53.2674, 41.7782, 438.5},
-    {-53.2674, 41.7782, 510.5},   {-42.8955, 54.9942, 510.5},
-    {-42.8955, 54.9942, -359},    {-53.2674, 41.7782, -359},
-    {-53.2674, 41.7782, -287},    {-42.8955, 54.9942, -287},
-    {-42.8955, 54.9942, -286.5},  {-53.2674, 41.7782, -286.5},
-    {-53.2674, 41.7782, -214.5},  {-42.8955, 54.9942, -214.5},
-    {-42.8955, 54.9942, -214},    {-53.2674, 41.7782, -214},
-    {-53.2674, 41.7782, -142},    {-42.8955, 54.9942, -142},
-    {-42.8955, 54.9942, -141.5},  {-53.2674, 41.7782, -141.5},
-    {-53.2674, 41.7782, -69.5},   {-42.8955, 54.9942, -69.5},
-    {-42.8955, 54.9942, -69},     {-53.2674, 41.7782, -69},
-    {-53.2674, 41.7782, 3},       {-42.8955, 54.9942, 3},
-    {-42.8955, 54.9942, 3.5},     {-53.2674, 41.7782, 3.5},
-    {-53.2674, 41.7782, 75.5},    {-42.8955, 54.9942, 75.5},
-    {-42.8955, 54.9942, 76},      {-53.2674, 41.7782, 76},
-    {-53.2674, 41.7782, 148},     {-42.8955, 54.9942, 148},
-    {-42.8955, 54.9942, 148.5},   {-53.2674, 41.7782, 148.5},
-    {-53.2674, 41.7782, 220.5},   {-42.8955, 54.9942, 220.5},
-    {-52.8002, 45.569, -504},     {-60.3944, 30.5835, -504},
-    {-60.3944, 30.5835, -432},    {-52.8002, 45.569, -432},
-    {-52.8002, 45.569, -431.5},   {-60.3944, 30.5835, -431.5},
-    {-60.3944, 30.5835, -359.5},  {-52.8002, 45.569, -359.5},
-    {-52.8002, 45.569, 221},      {-60.3944, 30.5835, 221},
-    {-60.3944, 30.5835, 293},     {-52.8002, 45.569, 293},
-    {-52.8002, 45.569, 293.5},    {-60.3944, 30.5835, 293.5},
-    {-60.3944, 30.5835, 365.5},   {-52.8002, 45.569, 365.5},
-    {-52.8002, 45.569, 366},      {-60.3944, 30.5835, 366},
-    {-60.3944, 30.5835, 438},     {-52.8002, 45.569, 438},
-    {-52.8002, 45.569, 438.5},    {-60.3944, 30.5835, 438.5},
-    {-60.3944, 30.5835, 510.5},   {-52.8002, 45.569, 510.5},
-    {-52.8002, 45.569, -359},     {-60.3944, 30.5835, -359},
-    {-60.3944, 30.5835, -287},    {-52.8002, 45.569, -287},
-    {-52.8002, 45.569, -286.5},   {-60.3944, 30.5835, -286.5},
-    {-60.3944, 30.5835, -214.5},  {-52.8002, 45.569, -214.5},
-    {-52.8002, 45.569, -214},     {-60.3944, 30.5835, -214},
-    {-60.3944, 30.5835, -142},    {-52.8002, 45.569, -142},
-    {-52.8002, 45.569, -141.5},   {-60.3944, 30.5835, -141.5},
-    {-60.3944, 30.5835, -69.5},   {-52.8002, 45.569, -69.5},
-    {-52.8002, 45.569, -69},      {-60.3944, 30.5835, -69},
-    {-60.3944, 30.5835, 3},       {-52.8002, 45.569, 3},
-    {-52.8002, 45.569, 3.5},      {-60.3944, 30.5835, 3.5},
-    {-60.3944, 30.5835, 75.5},    {-52.8002, 45.569, 75.5},
-    {-52.8002, 45.569, 76},       {-60.3944, 30.5835, 76},
-    {-60.3944, 30.5835, 148},     {-52.8002, 45.569, 148},
-    {-52.8002, 45.569, 148.5},    {-60.3944, 30.5835, 148.5},
-    {-60.3944, 30.5835, 220.5},   {-52.8002, 45.569, 220.5},
-    {-60.6757, 34.3926, -504},    {-65.2005, 18.2134, -504},
-    {-65.2005, 18.2134, -432},    {-60.6757, 34.3926, -432},
-    {-60.6757, 34.3926, -431.5},  {-65.2005, 18.2134, -431.5},
-    {-65.2005, 18.2134, -359.5},  {-60.6757, 34.3926, -359.5},
-    {-60.6757, 34.3926, 221},     {-65.2005, 18.2134, 221},
-    {-65.2005, 18.2134, 293},     {-60.6757, 34.3926, 293},
-    {-60.6757, 34.3926, 293.5},   {-65.2005, 18.2134, 293.5},
-    {-65.2005, 18.2134, 365.5},   {-60.6757, 34.3926, 365.5},
-    {-60.6757, 34.3926, 366},     {-65.2005, 18.2134, 366},
-    {-65.2005, 18.2134, 438},     {-60.6757, 34.3926, 438},
-    {-60.6757, 34.3926, 438.5},   {-65.2005, 18.2134, 438.5},
-    {-65.2005, 18.2134, 510.5},   {-60.6757, 34.3926, 510.5},
-    {-60.6757, 34.3926, -359},    {-65.2005, 18.2134, -359},
-    {-65.2005, 18.2134, -287},    {-60.6757, 34.3926, -287},
-    {-60.6757, 34.3926, -286.5},  {-65.2005, 18.2134, -286.5},
-    {-65.2005, 18.2134, -214.5},  {-60.6757, 34.3926, -214.5},
-    {-60.6757, 34.3926, -214},    {-65.2005, 18.2134, -214},
-    {-65.2005, 18.2134, -142},    {-60.6757, 34.3926, -142},
-    {-60.6757, 34.3926, -141.5},  {-65.2005, 18.2134, -141.5},
-    {-65.2005, 18.2134, -69.5},   {-60.6757, 34.3926, -69.5},
-    {-60.6757, 34.3926, -69},     {-65.2005, 18.2134, -69},
-    {-65.2005, 18.2134, 3},       {-60.6757, 34.3926, 3},
-    {-60.6757, 34.3926, 3.5},     {-65.2005, 18.2134, 3.5},
-    {-65.2005, 18.2134, 75.5},    {-60.6757, 34.3926, 75.5},
-    {-60.6757, 34.3926, 76},      {-65.2005, 18.2134, 76},
-    {-65.2005, 18.2134, 148},     {-60.6757, 34.3926, 148},
-    {-60.6757, 34.3926, 148.5},   {-65.2005, 18.2134, 148.5},
-    {-65.2005, 18.2134, 220.5},   {-60.6757, 34.3926, 220.5},
-    {-66.2195, 21.8946, -504},    {-67.5009, 5.1435, -504},
-    {-67.5009, 5.1435, -432},     {-66.2195, 21.8946, -432},
-    {-66.2195, 21.8946, -431.5},  {-67.5009, 5.1435, -431.5},
-    {-67.5009, 5.1435, -359.5},   {-66.2195, 21.8946, -359.5},
-    {-66.2195, 21.8946, 221},     {-67.5009, 5.1435, 221},
-    {-67.5009, 5.1435, 293},      {-66.2195, 21.8946, 293},
-    {-66.2195, 21.8946, 293.5},   {-67.5009, 5.1435, 293.5},
-    {-67.5009, 5.1435, 365.5},    {-66.2195, 21.8946, 365.5},
-    {-66.2195, 21.8946, 366},     {-67.5009, 5.1435, 366},
-    {-67.5009, 5.1435, 438},      {-66.2195, 21.8946, 438},
-    {-66.2195, 21.8946, 438.5},   {-67.5009, 5.1435, 438.5},
-    {-67.5009, 5.1435, 510.5},    {-66.2195, 21.8946, 510.5},
-    {-66.2195, 21.8946, -359},    {-67.5009, 5.1435, -359},
-    {-67.5009, 5.1435, -287},     {-66.2195, 21.8946, -287},
-    {-66.2195, 21.8946, -286.5},  {-67.5009, 5.1435, -286.5},
-    {-67.5009, 5.1435, -214.5},   {-66.2195, 21.8946, -214.5},
-    {-66.2195, 21.8946, -214},    {-67.5009, 5.1435, -214},
-    {-67.5009, 5.1435, -142},     {-66.2195, 21.8946, -142},
-    {-66.2195, 21.8946, -141.5},  {-67.5009, 5.1435, -141.5},
-    {-67.5009, 5.1435, -69.5},    {-66.2195, 21.8946, -69.5},
-    {-66.2195, 21.8946, -69},     {-67.5009, 5.1435, -69},
-    {-67.5009, 5.1435, 3},        {-66.2195, 21.8946, 3},
-    {-66.2195, 21.8946, 3.5},     {-67.5009, 5.1435, 3.5},
-    {-67.5009, 5.1435, 75.5},     {-66.2195, 21.8946, 75.5},
-    {-66.2195, 21.8946, 76},      {-67.5009, 5.1435, 76},
-    {-67.5009, 5.1435, 148},      {-66.2195, 21.8946, 148},
-    {-66.2195, 21.8946, 148.5},   {-67.5009, 5.1435, 148.5},
-    {-67.5009, 5.1435, 220.5},    {-66.2195, 21.8946, 220.5},
-    {-69.2185, 8.55507, -504},    {-67.2074, -8.12411, -504},
-    {-67.2074, -8.12411, -432},   {-69.2185, 8.55507, -432},
-    {-69.2185, 8.55507, -431.5},  {-67.2074, -8.12411, -431.5},
-    {-67.2074, -8.12411, -359.5}, {-69.2185, 8.55507, -359.5},
-    {-69.2185, 8.55507, 221},     {-67.2074, -8.12411, 221},
-    {-67.2074, -8.12411, 293},    {-69.2185, 8.55507, 293},
-    {-69.2185, 8.55507, 293.5},   {-67.2074, -8.12411, 293.5},
-    {-67.2074, -8.12411, 365.5},  {-69.2185, 8.55507, 365.5},
-    {-69.2185, 8.55507, 366},     {-67.2074, -8.12411, 366},
-    {-67.2074, -8.12411, 438},    {-69.2185, 8.55507, 438},
-    {-69.2185, 8.55507, 438.5},   {-67.2074, -8.12411, 438.5},
-    {-67.2074, -8.12411, 510.5},  {-69.2185, 8.55507, 510.5},
-    {-69.2185, 8.55507, -359},    {-67.2074, -8.12411, -359},
-    {-67.2074, -8.12411, -287},   {-69.2185, 8.55507, -287},
-    {-69.2185, 8.55507, -286.5},  {-67.2074, -8.12411, -286.5},
-    {-67.2074, -8.12411, -214.5}, {-69.2185, 8.55507, -214.5},
-    {-69.2185, 8.55507, -214},    {-67.2074, -8.12411, -214},
-    {-67.2074, -8.12411, -142},   {-69.2185, 8.55507, -142},
-    {-69.2185, 8.55507, -141.5},  {-67.2074, -8.12411, -141.5},
-    {-67.2074, -8.12411, -69.5},  {-69.2185, 8.55507, -69.5},
-    {-69.2185, 8.55507, -69},     {-67.2074, -8.12411, -69},
-    {-67.2074, -8.12411, 3},      {-69.2185, 8.55507, 3},
-    {-69.2185, 8.55507, 3.5},     {-67.2074, -8.12411, 3.5},
-    {-67.2074, -8.12411, 75.5},   {-69.2185, 8.55507, 75.5},
-    {-69.2185, 8.55507, 76},      {-67.2074, -8.12411, 76},
-    {-67.2074, -8.12411, 148},    {-69.2185, 8.55507, 148},
-    {-69.2185, 8.55507, 148.5},   {-67.2074, -8.12411, 148.5},
-    {-67.2074, -8.12411, 220.5},  {-69.2185, 8.55507, 220.5},
-    {-69.5575, -5.11317, -504},   {-64.3311, -21.0795, -504},
-    {-64.3311, -21.0795, -432},   {-69.5575, -5.11317, -432},
-    {-69.5575, -5.11317, -431.5}, {-64.3311, -21.0795, -431.5},
-    {-64.3311, -21.0795, -359.5}, {-69.5575, -5.11317, -359.5},
-    {-69.5575, -5.11317, 221},    {-64.3311, -21.0795, 221},
-    {-64.3311, -21.0795, 293},    {-69.5575, -5.11317, 293},
-    {-69.5575, -5.11317, 293.5},  {-64.3311, -21.0795, 293.5},
-    {-64.3311, -21.0795, 365.5},  {-69.5575, -5.11317, 365.5},
-    {-69.5575, -5.11317, 366},    {-64.3311, -21.0795, 366},
-    {-64.3311, -21.0795, 438},    {-69.5575, -5.11317, 438},
-    {-69.5575, -5.11317, 438.5},  {-64.3311, -21.0795, 438.5},
-    {-64.3311, -21.0795, 510.5},  {-69.5575, -5.11317, 510.5},
-    {-69.5575, -5.11317, -359},   {-64.3311, -21.0795, -359},
-    {-64.3311, -21.0795, -287},   {-69.5575, -5.11317, -287},
-    {-69.5575, -5.11317, -286.5}, {-64.3311, -21.0795, -286.5},
-    {-64.3311, -21.0795, -214.5}, {-69.5575, -5.11317, -214.5},
-    {-69.5575, -5.11317, -214},   {-64.3311, -21.0795, -214},
-    {-64.3311, -21.0795, -142},   {-69.5575, -5.11317, -142},
-    {-69.5575, -5.11317, -141.5}, {-64.3311, -21.0795, -141.5},
-    {-64.3311, -21.0795, -69.5},  {-69.5575, -5.11317, -69.5},
-    {-69.5575, -5.11317, -69},    {-64.3311, -21.0795, -69},
-    {-64.3311, -21.0795, 3},      {-69.5575, -5.11317, 3},
-    {-69.5575, -5.11317, 3.5},    {-64.3311, -21.0795, 3.5},
-    {-64.3311, -21.0795, 75.5},   {-69.5575, -5.11317, 75.5},
-    {-69.5575, -5.11317, 76},     {-64.3311, -21.0795, 76},
-    {-64.3311, -21.0795, 148},    {-69.5575, -5.11317, 148},
-    {-69.5575, -5.11317, 148.5},  {-64.3311, -21.0795, 148.5},
-    {-64.3311, -21.0795, 220.5},  {-69.5575, -5.11317, 220.5},
-    {-67.2235, -18.5849, -504},   {-58.9825, -33.2248, -504},
-    {-58.9825, -33.2248, -432},   {-67.2235, -18.5849, -432},
-    {-67.2235, -18.5849, -431.5}, {-58.9825, -33.2248, -431.5},
-    {-58.9825, -33.2248, -359.5}, {-67.2235, -18.5849, -359.5},
-    {-67.2235, -18.5849, 221},    {-58.9825, -33.2248, 221},
-    {-58.9825, -33.2248, 293},    {-67.2235, -18.5849, 293},
-    {-67.2235, -18.5849, 293.5},  {-58.9825, -33.2248, 293.5},
-    {-58.9825, -33.2248, 365.5},  {-67.2235, -18.5849, 365.5},
-    {-67.2235, -18.5849, 366},    {-58.9825, -33.2248, 366},
-    {-58.9825, -33.2248, 438},    {-67.2235, -18.5849, 438},
-    {-67.2235, -18.5849, 438.5},  {-58.9825, -33.2248, 438.5},
-    {-58.9825, -33.2248, 510.5},  {-67.2235, -18.5849, 510.5},
-    {-67.2235, -18.5849, -359},   {-58.9825, -33.2248, -359},
-    {-58.9825, -33.2248, -287},   {-67.2235, -18.5849, -287},
-    {-67.2235, -18.5849, -286.5}, {-58.9825, -33.2248, -286.5},
-    {-58.9825, -33.2248, -214.5}, {-67.2235, -18.5849, -214.5},
-    {-67.2235, -18.5849, -214},   {-58.9825, -33.2248, -214},
-    {-58.9825, -33.2248, -142},   {-67.2235, -18.5849, -142},
-    {-67.2235, -18.5849, -141.5}, {-58.9825, -33.2248, -141.5},
-    {-58.9825, -33.2248, -69.5},  {-67.2235, -18.5849, -69.5},
-    {-67.2235, -18.5849, -69},    {-58.9825, -33.2248, -69},
-    {-58.9825, -33.2248, 3},      {-67.2235, -18.5849, 3},
-    {-67.2235, -18.5849, 3.5},    {-58.9825, -33.2248, 3.5},
-    {-58.9825, -33.2248, 75.5},   {-67.2235, -18.5849, 75.5},
-    {-67.2235, -18.5849, 76},     {-58.9825, -33.2248, 76},
-    {-58.9825, -33.2248, 148},    {-67.2235, -18.5849, 148},
-    {-67.2235, -18.5849, 148.5},  {-58.9825, -33.2248, 148.5},
-    {-58.9825, -33.2248, 220.5},  {-67.2235, -18.5849, 220.5},
-    {-62.306, -31.3425, -504},    {-51.3674, -44.0934, -504},
-    {-51.3674, -44.0934, -432},   {-62.306, -31.3425, -432},
-    {-62.306, -31.3425, -431.5},  {-51.3674, -44.0934, -431.5},
-    {-51.3674, -44.0934, -359.5}, {-62.306, -31.3425, -359.5},
-    {-62.306, -31.3425, 221},     {-51.3674, -44.0934, 221},
-    {-51.3674, -44.0934, 293},    {-62.306, -31.3425, 293},
-    {-62.306, -31.3425, 293.5},   {-51.3674, -44.0934, 293.5},
-    {-51.3674, -44.0934, 365.5},  {-62.306, -31.3425, 365.5},
-    {-62.306, -31.3425, 366},     {-51.3674, -44.0934, 366},
-    {-51.3674, -44.0934, 438},    {-62.306, -31.3425, 438},
-    {-62.306, -31.3425, 438.5},   {-51.3674, -44.0934, 438.5},
-    {-51.3674, -44.0934, 510.5},  {-62.306, -31.3425, 510.5},
-    {-62.306, -31.3425, -359},    {-51.3674, -44.0934, -359},
-    {-51.3674, -44.0934, -287},   {-62.306, -31.3425, -287},
-    {-62.306, -31.3425, -286.5},  {-51.3674, -44.0934, -286.5},
-    {-51.3674, -44.0934, -214.5}, {-62.306, -31.3425, -214.5},
-    {-62.306, -31.3425, -214},    {-51.3674, -44.0934, -214},
-    {-51.3674, -44.0934, -142},   {-62.306, -31.3425, -142},
-    {-62.306, -31.3425, -141.5},  {-51.3674, -44.0934, -141.5},
-    {-51.3674, -44.0934, -69.5},  {-62.306, -31.3425, -69.5},
-    {-62.306, -31.3425, -69},     {-51.3674, -44.0934, -69},
-    {-51.3674, -44.0934, 3},      {-62.306, -31.3425, 3},
-    {-62.306, -31.3425, 3.5},     {-51.3674, -44.0934, 3.5},
-    {-51.3674, -44.0934, 75.5},   {-62.306, -31.3425, 75.5},
-    {-62.306, -31.3425, 76},      {-51.3674, -44.0934, 76},
-    {-51.3674, -44.0934, 148},    {-62.306, -31.3425, 148},
-    {-62.306, -31.3425, 148.5},   {-51.3674, -44.0934, 148.5},
-    {-51.3674, -44.0934, 220.5},  {-62.306, -31.3425, 220.5},
-    {67.2235, 18.5849, -504},     {58.9825, 33.2248, -504},
-    {58.9825, 33.2248, -432},     {67.2235, 18.5849, -432},
-    {67.2235, 18.5849, -431.5},   {58.9825, 33.2248, -431.5},
-    {58.9825, 33.2248, -359.5},   {67.2235, 18.5849, -359.5},
-    {67.2235, 18.5849, 221},      {58.9825, 33.2248, 221},
-    {58.9825, 33.2248, 293},      {67.2235, 18.5849, 293},
-    {67.2235, 18.5849, 293.5},    {58.9825, 33.2248, 293.5},
-    {58.9825, 33.2248, 365.5},    {67.2235, 18.5849, 365.5},
-    {67.2235, 18.5849, 366},      {58.9825, 33.2248, 366},
-    {58.9825, 33.2248, 438},      {67.2235, 18.5849, 438},
-    {67.2235, 18.5849, 438.5},    {58.9825, 33.2248, 438.5},
-    {58.9825, 33.2248, 510.5},    {67.2235, 18.5849, 510.5},
-    {67.2235, 18.5849, -359},     {58.9825, 33.2248, -359},
-    {58.9825, 33.2248, -287},     {67.2235, 18.5849, -287},
-    {67.2235, 18.5849, -286.5},   {58.9825, 33.2248, -286.5},
-    {58.9825, 33.2248, -214.5},   {67.2235, 18.5849, -214.5},
-    {67.2235, 18.5849, -214},     {58.9825, 33.2248, -214},
-    {58.9825, 33.2248, -142},     {67.2235, 18.5849, -142},
-    {67.2235, 18.5849, -141.5},   {58.9825, 33.2248, -141.5},
-    {58.9825, 33.2248, -69.5},    {67.2235, 18.5849, -69.5},
-    {67.2235, 18.5849, -69},      {58.9825, 33.2248, -69},
-    {58.9825, 33.2248, 3},        {67.2235, 18.5849, 3},
-    {67.2235, 18.5849, 3.5},      {58.9825, 33.2248, 3.5},
-    {58.9825, 33.2248, 75.5},     {67.2235, 18.5849, 75.5},
-    {67.2235, 18.5849, 76},       {58.9825, 33.2248, 76},
-    {58.9825, 33.2248, 148},      {67.2235, 18.5849, 148},
-    {67.2235, 18.5849, 148.5},    {58.9825, 33.2248, 148.5},
-    {58.9825, 33.2248, 220.5},    {67.2235, 18.5849, 220.5},
-    {-54.9942, -42.8955, -504},   {-41.7782, -53.2674, -504},
-    {-41.7782, -53.2674, -432},   {-54.9942, -42.8955, -432},
-    {-54.9942, -42.8955, -431.5}, {-41.7782, -53.2674, -431.5},
-    {-41.7782, -53.2674, -359.5}, {-54.9942, -42.8955, -359.5},
-    {-54.9942, -42.8955, 221},    {-41.7782, -53.2674, 221},
-    {-41.7782, -53.2674, 293},    {-54.9942, -42.8955, 293},
-    {-54.9942, -42.8955, 293.5},  {-41.7782, -53.2674, 293.5},
-    {-41.7782, -53.2674, 365.5},  {-54.9942, -42.8955, 365.5},
-    {-54.9942, -42.8955, 366},    {-41.7782, -53.2674, 366},
-    {-41.7782, -53.2674, 438},    {-54.9942, -42.8955, 438},
-    {-54.9942, -42.8955, 438.5},  {-41.7782, -53.2674, 438.5},
-    {-41.7782, -53.2674, 510.5},  {-54.9942, -42.8955, 510.5},
-    {-54.9942, -42.8955, -359},   {-41.7782, -53.2674, -359},
-    {-41.7782, -53.2674, -287},   {-54.9942, -42.8955, -287},
-    {-54.9942, -42.8955, -286.5}, {-41.7782, -53.2674, -286.5},
-    {-41.7782, -53.2674, -214.5}, {-54.9942, -42.8955, -214.5},
-    {-54.9942, -42.8955, -214},   {-41.7782, -53.2674, -214},
-    {-41.7782, -53.2674, -142},   {-54.9942, -42.8955, -142},
-    {-54.9942, -42.8955, -141.5}, {-41.7782, -53.2674, -141.5},
-    {-41.7782, -53.2674, -69.5},  {-54.9942, -42.8955, -69.5},
-    {-54.9942, -42.8955, -69},    {-41.7782, -53.2674, -69},
-    {-41.7782, -53.2674, 3},      {-54.9942, -42.8955, 3},
-    {-54.9942, -42.8955, 3.5},    {-41.7782, -53.2674, 3.5},
-    {-41.7782, -53.2674, 75.5},   {-54.9942, -42.8955, 75.5},
-    {-54.9942, -42.8955, 76},     {-41.7782, -53.2674, 76},
-    {-41.7782, -53.2674, 148},    {-54.9942, -42.8955, 148},
-    {-54.9942, -42.8955, 148.5},  {-41.7782, -53.2674, 148.5},
-    {-41.7782, -53.2674, 220.5},  {-54.9942, -42.8955, 220.5},
-    {-45.569, -52.8002, -504},    {-30.5835, -60.3944, -504},
-    {-30.5835, -60.3944, -432},   {-45.569, -52.8002, -432},
-    {-45.569, -52.8002, -431.5},  {-30.5835, -60.3944, -431.5},
-    {-30.5835, -60.3944, -359.5}, {-45.569, -52.8002, -359.5},
-    {-45.569, -52.8002, 221},     {-30.5835, -60.3944, 221},
-    {-30.5835, -60.3944, 293},    {-45.569, -52.8002, 293},
-    {-45.569, -52.8002, 293.5},   {-30.5835, -60.3944, 293.5},
-    {-30.5835, -60.3944, 365.5},  {-45.569, -52.8002, 365.5},
-    {-45.569, -52.8002, 366},     {-30.5835, -60.3944, 366},
-    {-30.5835, -60.3944, 438},    {-45.569, -52.8002, 438},
-    {-45.569, -52.8002, 438.5},   {-30.5835, -60.3944, 438.5},
-    {-30.5835, -60.3944, 510.5},  {-45.569, -52.8002, 510.5},
-    {-45.569, -52.8002, -359},    {-30.5835, -60.3944, -359},
-    {-30.5835, -60.3944, -287},   {-45.569, -52.8002, -287},
-    {-45.569, -52.8002, -286.5},  {-30.5835, -60.3944, -286.5},
-    {-30.5835, -60.3944, -214.5}, {-45.569, -52.8002, -214.5},
-    {-45.569, -52.8002, -214},    {-30.5835, -60.3944, -214},
-    {-30.5835, -60.3944, -142},   {-45.569, -52.8002, -142},
-    {-45.569, -52.8002, -141.5},  {-30.5835, -60.3944, -141.5},
-    {-30.5835, -60.3944, -69.5},  {-45.569, -52.8002, -69.5},
-    {-45.569, -52.8002, -69},     {-30.5835, -60.3944, -69},
-    {-30.5835, -60.3944, 3},      {-45.569, -52.8002, 3},
-    {-45.569, -52.8002, 3.5},     {-30.5835, -60.3944, 3.5},
-    {-30.5835, -60.3944, 75.5},   {-45.569, -52.8002, 75.5},
-    {-45.569, -52.8002, 76},      {-30.5835, -60.3944, 76},
-    {-30.5835, -60.3944, 148},    {-45.569, -52.8002, 148},
-    {-45.569, -52.8002, 148.5},   {-30.5835, -60.3944, 148.5},
-    {-30.5835, -60.3944, 220.5},  {-45.569, -52.8002, 220.5},
-    {-34.3926, -60.6757, -504},   {-18.2134, -65.2005, -504},
-    {-18.2134, -65.2005, -432},   {-34.3926, -60.6757, -432},
-    {-34.3926, -60.6757, -431.5}, {-18.2134, -65.2005, -431.5},
-    {-18.2134, -65.2005, -359.5}, {-34.3926, -60.6757, -359.5},
-    {-34.3926, -60.6757, 221},    {-18.2134, -65.2005, 221},
-    {-18.2134, -65.2005, 293},    {-34.3926, -60.6757, 293},
-    {-34.3926, -60.6757, 293.5},  {-18.2134, -65.2005, 293.5},
-    {-18.2134, -65.2005, 365.5},  {-34.3926, -60.6757, 365.5},
-    {-34.3926, -60.6757, 366},    {-18.2134, -65.2005, 366},
-    {-18.2134, -65.2005, 438},    {-34.3926, -60.6757, 438},
-    {-34.3926, -60.6757, 438.5},  {-18.2134, -65.2005, 438.5},
-    {-18.2134, -65.2005, 510.5},  {-34.3926, -60.6757, 510.5},
-    {-34.3926, -60.6757, -359},   {-18.2134, -65.2005, -359},
-    {-18.2134, -65.2005, -287},   {-34.3926, -60.6757, -287},
-    {-34.3926, -60.6757, -286.5}, {-18.2134, -65.2005, -286.5},
-    {-18.2134, -65.2005, -214.5}, {-34.3926, -60.6757, -214.5},
-    {-34.3926, -60.6757, -214},   {-18.2134, -65.2005, -214},
-    {-18.2134, -65.2005, -142},   {-34.3926, -60.6757, -142},
-    {-34.3926, -60.6757, -141.5}, {-18.2134, -65.2005, -141.5},
-    {-18.2134, -65.2005, -69.5},  {-34.3926, -60.6757, -69.5},
-    {-34.3926, -60.6757, -69},    {-18.2134, -65.2005, -69},
-    {-18.2134, -65.2005, 3},      {-34.3926, -60.6757, 3},
-    {-34.3926, -60.6757, 3.5},    {-18.2134, -65.2005, 3.5},
-    {-18.2134, -65.2005, 75.5},   {-34.3926, -60.6757, 75.5},
-    {-34.3926, -60.6757, 76},     {-18.2134, -65.2005, 76},
-    {-18.2134, -65.2005, 148},    {-34.3926, -60.6757, 148},
-    {-34.3926, -60.6757, 148.5},  {-18.2134, -65.2005, 148.5},
-    {-18.2134, -65.2005, 220.5},  {-34.3926, -60.6757, 220.5},
-    {-21.8946, -66.2195, -504},   {-5.1435, -67.5009, -504},
-    {-5.1435, -67.5009, -432},    {-21.8946, -66.2195, -432},
-    {-21.8946, -66.2195, -431.5}, {-5.1435, -67.5009, -431.5},
-    {-5.1435, -67.5009, -359.5},  {-21.8946, -66.2195, -359.5},
-    {-21.8946, -66.2195, 221},    {-5.1435, -67.5009, 221},
-    {-5.1435, -67.5009, 293},     {-21.8946, -66.2195, 293},
-    {-21.8946, -66.2195, 293.5},  {-5.1435, -67.5009, 293.5},
-    {-5.1435, -67.5009, 365.5},   {-21.8946, -66.2195, 365.5},
-    {-21.8946, -66.2195, 366},    {-5.1435, -67.5009, 366},
-    {-5.1435, -67.5009, 438},     {-21.8946, -66.2195, 438},
-    {-21.8946, -66.2195, 438.5},  {-5.1435, -67.5009, 438.5},
-    {-5.1435, -67.5009, 510.5},   {-21.8946, -66.2195, 510.5},
-    {-21.8946, -66.2195, -359},   {-5.1435, -67.5009, -359},
-    {-5.1435, -67.5009, -287},    {-21.8946, -66.2195, -287},
-    {-21.8946, -66.2195, -286.5}, {-5.1435, -67.5009, -286.5},
-    {-5.1435, -67.5009, -214.5},  {-21.8946, -66.2195, -214.5},
-    {-21.8946, -66.2195, -214},   {-5.1435, -67.5009, -214},
-    {-5.1435, -67.5009, -142},    {-21.8946, -66.2195, -142},
-    {-21.8946, -66.2195, -141.5}, {-5.1435, -67.5009, -141.5},
-    {-5.1435, -67.5009, -69.5},   {-21.8946, -66.2195, -69.5},
-    {-21.8946, -66.2195, -69},    {-5.1435, -67.5009, -69},
-    {-5.1435, -67.5009, 3},       {-21.8946, -66.2195, 3},
-    {-21.8946, -66.2195, 3.5},    {-5.1435, -67.5009, 3.5},
-    {-5.1435, -67.5009, 75.5},    {-21.8946, -66.2195, 75.5},
-    {-21.8946, -66.2195, 76},     {-5.1435, -67.5009, 76},
-    {-5.1435, -67.5009, 148},     {-21.8946, -66.2195, 148},
-    {-21.8946, -66.2195, 148.5},  {-5.1435, -67.5009, 148.5},
-    {-5.1435, -67.5009, 220.5},   {-21.8946, -66.2195, 220.5},
-    {-8.55507, -69.2185, -504},   {8.12411, -67.2074, -504},
-    {8.12411, -67.2074, -432},    {-8.55507, -69.2185, -432},
-    {-8.55507, -69.2185, -431.5}, {8.12411, -67.2074, -431.5},
-    {8.12411, -67.2074, -359.5},  {-8.55507, -69.2185, -359.5},
-    {-8.55507, -69.2185, 221},    {8.12411, -67.2074, 221},
-    {8.12411, -67.2074, 293},     {-8.55507, -69.2185, 293},
-    {-8.55507, -69.2185, 293.5},  {8.12411, -67.2074, 293.5},
-    {8.12411, -67.2074, 365.5},   {-8.55507, -69.2185, 365.5},
-    {-8.55507, -69.2185, 366},    {8.12411, -67.2074, 366},
-    {8.12411, -67.2074, 438},     {-8.55507, -69.2185, 438},
-    {-8.55507, -69.2185, 438.5},  {8.12411, -67.2074, 438.5},
-    {8.12411, -67.2074, 510.5},   {-8.55507, -69.2185, 510.5},
-    {-8.55507, -69.2185, -359},   {8.12411, -67.2074, -359},
-    {8.12411, -67.2074, -287},    {-8.55507, -69.2185, -287},
-    {-8.55507, -69.2185, -286.5}, {8.12411, -67.2074, -286.5},
-    {8.12411, -67.2074, -214.5},  {-8.55507, -69.2185, -214.5},
-    {-8.55507, -69.2185, -214},   {8.12411, -67.2074, -214},
-    {8.12411, -67.2074, -142},    {-8.55507, -69.2185, -142},
-    {-8.55507, -69.2185, -141.5}, {8.12411, -67.2074, -141.5},
-    {8.12411, -67.2074, -69.5},   {-8.55507, -69.2185, -69.5},
-    {-8.55507, -69.2185, -69},    {8.12411, -67.2074, -69},
-    {8.12411, -67.2074, 3},       {-8.55507, -69.2185, 3},
-    {-8.55507, -69.2185, 3.5},    {8.12411, -67.2074, 3.5},
-    {8.12411, -67.2074, 75.5},    {-8.55507, -69.2185, 75.5},
-    {-8.55507, -69.2185, 76},     {8.12411, -67.2074, 76},
-    {8.12411, -67.2074, 148},     {-8.55507, -69.2185, 148},
-    {-8.55507, -69.2185, 148.5},  {8.12411, -67.2074, 148.5},
-    {8.12411, -67.2074, 220.5},   {-8.55507, -69.2185, 220.5},
-    {5.11317, -69.5575, -504},    {21.0795, -64.3311, -504},
-    {21.0795, -64.3311, -432},    {5.11317, -69.5575, -432},
-    {5.11317, -69.5575, -431.5},  {21.0795, -64.3311, -431.5},
-    {21.0795, -64.3311, -359.5},  {5.11317, -69.5575, -359.5},
-    {5.11317, -69.5575, 221},     {21.0795, -64.3311, 221},
-    {21.0795, -64.3311, 293},     {5.11317, -69.5575, 293},
-    {5.11317, -69.5575, 293.5},   {21.0795, -64.3311, 293.5},
-    {21.0795, -64.3311, 365.5},   {5.11317, -69.5575, 365.5},
-    {5.11317, -69.5575, 366},     {21.0795, -64.3311, 366},
-    {21.0795, -64.3311, 438},     {5.11317, -69.5575, 438},
-    {5.11317, -69.5575, 438.5},   {21.0795, -64.3311, 438.5},
-    {21.0795, -64.3311, 510.5},   {5.11317, -69.5575, 510.5},
-    {5.11317, -69.5575, -359},    {21.0795, -64.3311, -359},
-    {21.0795, -64.3311, -287},    {5.11317, -69.5575, -287},
-    {5.11317, -69.5575, -286.5},  {21.0795, -64.3311, -286.5},
-    {21.0795, -64.3311, -214.5},  {5.11317, -69.5575, -214.5},
-    {5.11317, -69.5575, -214},    {21.0795, -64.3311, -214},
-    {21.0795, -64.3311, -142},    {5.11317, -69.5575, -142},
-    {5.11317, -69.5575, -141.5},  {21.0795, -64.3311, -141.5},
-    {21.0795, -64.3311, -69.5},   {5.11317, -69.5575, -69.5},
-    {5.11317, -69.5575, -69},     {21.0795, -64.3311, -69},
-    {21.0795, -64.3311, 3},       {5.11317, -69.5575, 3},
-    {5.11317, -69.5575, 3.5},     {21.0795, -64.3311, 3.5},
-    {21.0795, -64.3311, 75.5},    {5.11317, -69.5575, 75.5},
-    {5.11317, -69.5575, 76},      {21.0795, -64.3311, 76},
-    {21.0795, -64.3311, 148},     {5.11317, -69.5575, 148},
-    {5.11317, -69.5575, 148.5},   {21.0795, -64.3311, 148.5},
-    {21.0795, -64.3311, 220.5},   {5.11317, -69.5575, 220.5},
-    {18.5849, -67.2235, -504},    {33.2248, -58.9825, -504},
-    {33.2248, -58.9825, -432},    {18.5849, -67.2235, -432},
-    {18.5849, -67.2235, -431.5},  {33.2248, -58.9825, -431.5},
-    {33.2248, -58.9825, -359.5},  {18.5849, -67.2235, -359.5},
-    {18.5849, -67.2235, 221},     {33.2248, -58.9825, 221},
-    {33.2248, -58.9825, 293},     {18.5849, -67.2235, 293},
-    {18.5849, -67.2235, 293.5},   {33.2248, -58.9825, 293.5},
-    {33.2248, -58.9825, 365.5},   {18.5849, -67.2235, 365.5},
-    {18.5849, -67.2235, 366},     {33.2248, -58.9825, 366},
-    {33.2248, -58.9825, 438},     {18.5849, -67.2235, 438},
-    {18.5849, -67.2235, 438.5},   {33.2248, -58.9825, 438.5},
-    {33.2248, -58.9825, 510.5},   {18.5849, -67.2235, 510.5},
-    {18.5849, -67.2235, -359},    {33.2248, -58.9825, -359},
-    {33.2248, -58.9825, -287},    {18.5849, -67.2235, -287},
-    {18.5849, -67.2235, -286.5},  {33.2248, -58.9825, -286.5},
-    {33.2248, -58.9825, -214.5},  {18.5849, -67.2235, -214.5},
-    {18.5849, -67.2235, -214},    {33.2248, -58.9825, -214},
-    {33.2248, -58.9825, -142},    {18.5849, -67.2235, -142},
-    {18.5849, -67.2235, -141.5},  {33.2248, -58.9825, -141.5},
-    {33.2248, -58.9825, -69.5},   {18.5849, -67.2235, -69.5},
-    {18.5849, -67.2235, -69},     {33.2248, -58.9825, -69},
-    {33.2248, -58.9825, 3},       {18.5849, -67.2235, 3},
-    {18.5849, -67.2235, 3.5},     {33.2248, -58.9825, 3.5},
-    {33.2248, -58.9825, 75.5},    {18.5849, -67.2235, 75.5},
-    {18.5849, -67.2235, 76},      {33.2248, -58.9825, 76},
-    {33.2248, -58.9825, 148},     {18.5849, -67.2235, 148},
-    {18.5849, -67.2235, 148.5},   {33.2248, -58.9825, 148.5},
-    {33.2248, -58.9825, 220.5},   {18.5849, -67.2235, 220.5},
-    {31.3425, -62.306, -504},     {44.0934, -51.3674, -504},
-    {44.0934, -51.3674, -432},    {31.3425, -62.306, -432},
-    {31.3425, -62.306, -431.5},   {44.0934, -51.3674, -431.5},
-    {44.0934, -51.3674, -359.5},  {31.3425, -62.306, -359.5},
-    {31.3425, -62.306, 221},      {44.0934, -51.3674, 221},
-    {44.0934, -51.3674, 293},     {31.3425, -62.306, 293},
-    {31.3425, -62.306, 293.5},    {44.0934, -51.3674, 293.5},
-    {44.0934, -51.3674, 365.5},   {31.3425, -62.306, 365.5},
-    {31.3425, -62.306, 366},      {44.0934, -51.3674, 366},
-    {44.0934, -51.3674, 438},     {31.3425, -62.306, 438},
-    {31.3425, -62.306, 438.5},    {44.0934, -51.3674, 438.5},
-    {44.0934, -51.3674, 510.5},   {31.3425, -62.306, 510.5},
-    {31.3425, -62.306, -359},     {44.0934, -51.3674, -359},
-    {44.0934, -51.3674, -287},    {31.3425, -62.306, -287},
-    {31.3425, -62.306, -286.5},   {44.0934, -51.3674, -286.5},
-    {44.0934, -51.3674, -214.5},  {31.3425, -62.306, -214.5},
-    {31.3425, -62.306, -214},     {44.0934, -51.3674, -214},
-    {44.0934, -51.3674, -142},    {31.3425, -62.306, -142},
-    {31.3425, -62.306, -141.5},   {44.0934, -51.3674, -141.5},
-    {44.0934, -51.3674, -69.5},   {31.3425, -62.306, -69.5},
-    {31.3425, -62.306, -69},      {44.0934, -51.3674, -69},
-    {44.0934, -51.3674, 3},       {31.3425, -62.306, 3},
-    {31.3425, -62.306, 3.5},      {44.0934, -51.3674, 3.5},
-    {44.0934, -51.3674, 75.5},    {31.3425, -62.306, 75.5},
-    {31.3425, -62.306, 76},       {44.0934, -51.3674, 76},
-    {44.0934, -51.3674, 148},     {31.3425, -62.306, 148},
-    {31.3425, -62.306, 148.5},    {44.0934, -51.3674, 148.5},
-    {44.0934, -51.3674, 220.5},   {31.3425, -62.306, 220.5},
-    {42.8955, -54.9942, -504},    {53.2674, -41.7782, -504},
-    {53.2674, -41.7782, -432},    {42.8955, -54.9942, -432},
-    {42.8955, -54.9942, -431.5},  {53.2674, -41.7782, -431.5},
-    {53.2674, -41.7782, -359.5},  {42.8955, -54.9942, -359.5},
-    {42.8955, -54.9942, 221},     {53.2674, -41.7782, 221},
-    {53.2674, -41.7782, 293},     {42.8955, -54.9942, 293},
-    {42.8955, -54.9942, 293.5},   {53.2674, -41.7782, 293.5},
-    {53.2674, -41.7782, 365.5},   {42.8955, -54.9942, 365.5},
-    {42.8955, -54.9942, 366},     {53.2674, -41.7782, 366},
-    {53.2674, -41.7782, 438},     {42.8955, -54.9942, 438},
-    {42.8955, -54.9942, 438.5},   {53.2674, -41.7782, 438.5},
-    {53.2674, -41.7782, 510.5},   {42.8955, -54.9942, 510.5},
-    {42.8955, -54.9942, -359},    {53.2674, -41.7782, -359},
-    {53.2674, -41.7782, -287},    {42.8955, -54.9942, -287},
-    {42.8955, -54.9942, -286.5},  {53.2674, -41.7782, -286.5},
-    {53.2674, -41.7782, -214.5},  {42.8955, -54.9942, -214.5},
-    {42.8955, -54.9942, -214},    {53.2674, -41.7782, -214},
-    {53.2674, -41.7782, -142},    {42.8955, -54.9942, -142},
-    {42.8955, -54.9942, -141.5},  {53.2674, -41.7782, -141.5},
-    {53.2674, -41.7782, -69.5},   {42.8955, -54.9942, -69.5},
-    {42.8955, -54.9942, -69},     {53.2674, -41.7782, -69},
-    {53.2674, -41.7782, 3},       {42.8955, -54.9942, 3},
-    {42.8955, -54.9942, 3.5},     {53.2674, -41.7782, 3.5},
-    {53.2674, -41.7782, 75.5},    {42.8955, -54.9942, 75.5},
-    {42.8955, -54.9942, 76},      {53.2674, -41.7782, 76},
-    {53.2674, -41.7782, 148},     {42.8955, -54.9942, 148},
-    {42.8955, -54.9942, 148.5},   {53.2674, -41.7782, 148.5},
-    {53.2674, -41.7782, 220.5},   {42.8955, -54.9942, 220.5},
-    {52.8002, -45.569, -504},     {60.3944, -30.5835, -504},
-    {60.3944, -30.5835, -432},    {52.8002, -45.569, -432},
-    {52.8002, -45.569, -431.5},   {60.3944, -30.5835, -431.5},
-    {60.3944, -30.5835, -359.5},  {52.8002, -45.569, -359.5},
-    {52.8002, -45.569, 221},      {60.3944, -30.5835, 221},
-    {60.3944, -30.5835, 293},     {52.8002, -45.569, 293},
-    {52.8002, -45.569, 293.5},    {60.3944, -30.5835, 293.5},
-    {60.3944, -30.5835, 365.5},   {52.8002, -45.569, 365.5},
-    {52.8002, -45.569, 366},      {60.3944, -30.5835, 366},
-    {60.3944, -30.5835, 438},     {52.8002, -45.569, 438},
-    {52.8002, -45.569, 438.5},    {60.3944, -30.5835, 438.5},
-    {60.3944, -30.5835, 510.5},   {52.8002, -45.569, 510.5},
-    {52.8002, -45.569, -359},     {60.3944, -30.5835, -359},
-    {60.3944, -30.5835, -287},    {52.8002, -45.569, -287},
-    {52.8002, -45.569, -286.5},   {60.3944, -30.5835, -286.5},
-    {60.3944, -30.5835, -214.5},  {52.8002, -45.569, -214.5},
-    {52.8002, -45.569, -214},     {60.3944, -30.5835, -214},
-    {60.3944, -30.5835, -142},    {52.8002, -45.569, -142},
-    {52.8002, -45.569, -141.5},   {60.3944, -30.5835, -141.5},
-    {60.3944, -30.5835, -69.5},   {52.8002, -45.569, -69.5},
-    {52.8002, -45.569, -69},      {60.3944, -30.5835, -69},
-    {60.3944, -30.5835, 3},       {52.8002, -45.569, 3},
-    {52.8002, -45.569, 3.5},      {60.3944, -30.5835, 3.5},
-    {60.3944, -30.5835, 75.5},    {52.8002, -45.569, 75.5},
-    {52.8002, -45.569, 76},       {60.3944, -30.5835, 76},
-    {60.3944, -30.5835, 148},     {52.8002, -45.569, 148},
-    {52.8002, -45.569, 148.5},    {60.3944, -30.5835, 148.5},
-    {60.3944, -30.5835, 220.5},   {52.8002, -45.569, 220.5},
-    {62.306, 31.3425, -504},      {51.3674, 44.0934, -504},
-    {51.3674, 44.0934, -432},     {62.306, 31.3425, -432},
-    {62.306, 31.3425, -431.5},    {51.3674, 44.0934, -431.5},
-    {51.3674, 44.0934, -359.5},   {62.306, 31.3425, -359.5},
-    {62.306, 31.3425, 221},       {51.3674, 44.0934, 221},
-    {51.3674, 44.0934, 293},      {62.306, 31.3425, 293},
-    {62.306, 31.3425, 293.5},     {51.3674, 44.0934, 293.5},
-    {51.3674, 44.0934, 365.5},    {62.306, 31.3425, 365.5},
-    {62.306, 31.3425, 366},       {51.3674, 44.0934, 366},
-    {51.3674, 44.0934, 438},      {62.306, 31.3425, 438},
-    {62.306, 31.3425, 438.5},     {51.3674, 44.0934, 438.5},
-    {51.3674, 44.0934, 510.5},    {62.306, 31.3425, 510.5},
-    {62.306, 31.3425, -359},      {51.3674, 44.0934, -359},
-    {51.3674, 44.0934, -287},     {62.306, 31.3425, -287},
-    {62.306, 31.3425, -286.5},    {51.3674, 44.0934, -286.5},
-    {51.3674, 44.0934, -214.5},   {62.306, 31.3425, -214.5},
-    {62.306, 31.3425, -214},      {51.3674, 44.0934, -214},
-    {51.3674, 44.0934, -142},     {62.306, 31.3425, -142},
-    {62.306, 31.3425, -141.5},    {51.3674, 44.0934, -141.5},
-    {51.3674, 44.0934, -69.5},    {62.306, 31.3425, -69.5},
-    {62.306, 31.3425, -69},       {51.3674, 44.0934, -69},
-    {51.3674, 44.0934, 3},        {62.306, 31.3425, 3},
-    {62.306, 31.3425, 3.5},       {51.3674, 44.0934, 3.5},
-    {51.3674, 44.0934, 75.5},     {62.306, 31.3425, 75.5},
-    {62.306, 31.3425, 76},        {51.3674, 44.0934, 76},
-    {51.3674, 44.0934, 148},      {62.306, 31.3425, 148},
-    {62.306, 31.3425, 148.5},     {51.3674, 44.0934, 148.5},
-    {51.3674, 44.0934, 220.5},    {62.306, 31.3425, 220.5},
-    {60.6757, -34.3926, -504},    {65.2005, -18.2134, -504},
-    {65.2005, -18.2134, -432},    {60.6757, -34.3926, -432},
-    {60.6757, -34.3926, -431.5},  {65.2005, -18.2134, -431.5},
-    {65.2005, -18.2134, -359.5},  {60.6757, -34.3926, -359.5},
-    {60.6757, -34.3926, 221},     {65.2005, -18.2134, 221},
-    {65.2005, -18.2134, 293},     {60.6757, -34.3926, 293},
-    {60.6757, -34.3926, 293.5},   {65.2005, -18.2134, 293.5},
-    {65.2005, -18.2134, 365.5},   {60.6757, -34.3926, 365.5},
-    {60.6757, -34.3926, 366},     {65.2005, -18.2134, 366},
-    {65.2005, -18.2134, 438},     {60.6757, -34.3926, 438},
-    {60.6757, -34.3926, 438.5},   {65.2005, -18.2134, 438.5},
-    {65.2005, -18.2134, 510.5},   {60.6757, -34.3926, 510.5},
-    {60.6757, -34.3926, -359},    {65.2005, -18.2134, -359},
-    {65.2005, -18.2134, -287},    {60.6757, -34.3926, -287},
-    {60.6757, -34.3926, -286.5},  {65.2005, -18.2134, -286.5},
-    {65.2005, -18.2134, -214.5},  {60.6757, -34.3926, -214.5},
-    {60.6757, -34.3926, -214},    {65.2005, -18.2134, -214},
-    {65.2005, -18.2134, -142},    {60.6757, -34.3926, -142},
-    {60.6757, -34.3926, -141.5},  {65.2005, -18.2134, -141.5},
-    {65.2005, -18.2134, -69.5},   {60.6757, -34.3926, -69.5},
-    {60.6757, -34.3926, -69},     {65.2005, -18.2134, -69},
-    {65.2005, -18.2134, 3},       {60.6757, -34.3926, 3},
-    {60.6757, -34.3926, 3.5},     {65.2005, -18.2134, 3.5},
-    {65.2005, -18.2134, 75.5},    {60.6757, -34.3926, 75.5},
-    {60.6757, -34.3926, 76},      {65.2005, -18.2134, 76},
-    {65.2005, -18.2134, 148},     {60.6757, -34.3926, 148},
-    {60.6757, -34.3926, 148.5},   {65.2005, -18.2134, 148.5},
-    {65.2005, -18.2134, 220.5},   {60.6757, -34.3926, 220.5},
-    {66.2195, -21.8946, -504},    {67.5009, -5.1435, -504},
-    {67.5009, -5.1435, -432},     {66.2195, -21.8946, -432},
-    {66.2195, -21.8946, -431.5},  {67.5009, -5.1435, -431.5},
-    {67.5009, -5.1435, -359.5},   {66.2195, -21.8946, -359.5},
-    {66.2195, -21.8946, 221},     {67.5009, -5.1435, 221},
-    {67.5009, -5.1435, 293},      {66.2195, -21.8946, 293},
-    {66.2195, -21.8946, 293.5},   {67.5009, -5.1435, 293.5},
-    {67.5009, -5.1435, 365.5},    {66.2195, -21.8946, 365.5},
-    {66.2195, -21.8946, 366},     {67.5009, -5.1435, 366},
-    {67.5009, -5.1435, 438},      {66.2195, -21.8946, 438},
-    {66.2195, -21.8946, 438.5},   {67.5009, -5.1435, 438.5},
-    {67.5009, -5.1435, 510.5},    {66.2195, -21.8946, 510.5},
-    {66.2195, -21.8946, -359},    {67.5009, -5.1435, -359},
-    {67.5009, -5.1435, -287},     {66.2195, -21.8946, -287},
-    {66.2195, -21.8946, -286.5},  {67.5009, -5.1435, -286.5},
-    {67.5009, -5.1435, -214.5},   {66.2195, -21.8946, -214.5},
-    {66.2195, -21.8946, -214},    {67.5009, -5.1435, -214},
-    {67.5009, -5.1435, -142},     {66.2195, -21.8946, -142},
-    {66.2195, -21.8946, -141.5},  {67.5009, -5.1435, -141.5},
-    {67.5009, -5.1435, -69.5},    {66.2195, -21.8946, -69.5},
-    {66.2195, -21.8946, -69},     {67.5009, -5.1435, -69},
-    {67.5009, -5.1435, 3},        {66.2195, -21.8946, 3},
-    {66.2195, -21.8946, 3.5},     {67.5009, -5.1435, 3.5},
-    {67.5009, -5.1435, 75.5},     {66.2195, -21.8946, 75.5},
-    {66.2195, -21.8946, 76},      {67.5009, -5.1435, 76},
-    {67.5009, -5.1435, 148},      {66.2195, -21.8946, 148},
-    {66.2195, -21.8946, 148.5},   {67.5009, -5.1435, 148.5},
-    {67.5009, -5.1435, 220.5},    {66.2195, -21.8946, 220.5},
-    {54.9942, 42.8955, -504},     {41.7782, 53.2674, -504},
-    {41.7782, 53.2674, -432},     {54.9942, 42.8955, -432},
-    {54.9942, 42.8955, -431.5},   {41.7782, 53.2674, -431.5},
-    {41.7782, 53.2674, -359.5},   {54.9942, 42.8955, -359.5},
-    {54.9942, 42.8955, 221},      {41.7782, 53.2674, 221},
-    {41.7782, 53.2674, 293},      {54.9942, 42.8955, 293},
-    {54.9942, 42.8955, 293.5},    {41.7782, 53.2674, 293.5},
-    {41.7782, 53.2674, 365.5},    {54.9942, 42.8955, 365.5},
-    {54.9942, 42.8955, 366},      {41.7782, 53.2674, 366},
-    {41.7782, 53.2674, 438},      {54.9942, 42.8955, 438},
-    {54.9942, 42.8955, 438.5},    {41.7782, 53.2674, 438.5},
-    {41.7782, 53.2674, 510.5},    {54.9942, 42.8955, 510.5},
-    {54.9942, 42.8955, -359},     {41.7782, 53.2674, -359},
-    {41.7782, 53.2674, -287},     {54.9942, 42.8955, -287},
-    {54.9942, 42.8955, -286.5},   {41.7782, 53.2674, -286.5},
-    {41.7782, 53.2674, -214.5},   {54.9942, 42.8955, -214.5},
-    {54.9942, 42.8955, -214},     {41.7782, 53.2674, -214},
-    {41.7782, 53.2674, -142},     {54.9942, 42.8955, -142},
-    {54.9942, 42.8955, -141.5},   {41.7782, 53.2674, -141.5},
-    {41.7782, 53.2674, -69.5},    {54.9942, 42.8955, -69.5},
-    {54.9942, 42.8955, -69},      {41.7782, 53.2674, -69},
-    {41.7782, 53.2674, 3},        {54.9942, 42.8955, 3},
-    {54.9942, 42.8955, 3.5},      {41.7782, 53.2674, 3.5},
-    {41.7782, 53.2674, 75.5},     {54.9942, 42.8955, 75.5},
-    {54.9942, 42.8955, 76},       {41.7782, 53.2674, 76},
-    {41.7782, 53.2674, 148},      {54.9942, 42.8955, 148},
-    {54.9942, 42.8955, 148.5},    {41.7782, 53.2674, 148.5},
-    {41.7782, 53.2674, 220.5},    {54.9942, 42.8955, 220.5},
-    {45.569, 52.8002, -504},      {30.5835, 60.3944, -504},
-    {30.5835, 60.3944, -432},     {45.569, 52.8002, -432},
-    {45.569, 52.8002, -431.5},    {30.5835, 60.3944, -431.5},
-    {30.5835, 60.3944, -359.5},   {45.569, 52.8002, -359.5},
-    {45.569, 52.8002, 221},       {30.5835, 60.3944, 221},
-    {30.5835, 60.3944, 293},      {45.569, 52.8002, 293},
-    {45.569, 52.8002, 293.5},     {30.5835, 60.3944, 293.5},
-    {30.5835, 60.3944, 365.5},    {45.569, 52.8002, 365.5},
-    {45.569, 52.8002, 366},       {30.5835, 60.3944, 366},
-    {30.5835, 60.3944, 438},      {45.569, 52.8002, 438},
-    {45.569, 52.8002, 438.5},     {30.5835, 60.3944, 438.5},
-    {30.5835, 60.3944, 510.5},    {45.569, 52.8002, 510.5},
-    {45.569, 52.8002, -359},      {30.5835, 60.3944, -359},
-    {30.5835, 60.3944, -287},     {45.569, 52.8002, -287},
-    {45.569, 52.8002, -286.5},    {30.5835, 60.3944, -286.5},
-    {30.5835, 60.3944, -214.5},   {45.569, 52.8002, -214.5},
-    {45.569, 52.8002, -214},      {30.5835, 60.3944, -214},
-    {30.5835, 60.3944, -142},     {45.569, 52.8002, -142},
-    {45.569, 52.8002, -141.5},    {30.5835, 60.3944, -141.5},
-    {30.5835, 60.3944, -69.5},    {45.569, 52.8002, -69.5},
-    {45.569, 52.8002, -69},       {30.5835, 60.3944, -69},
-    {30.5835, 60.3944, 3},        {45.569, 52.8002, 3},
-    {45.569, 52.8002, 3.5},       {30.5835, 60.3944, 3.5},
-    {30.5835, 60.3944, 75.5},     {45.569, 52.8002, 75.5},
-    {45.569, 52.8002, 76},        {30.5835, 60.3944, 76},
-    {30.5835, 60.3944, 148},      {45.569, 52.8002, 148},
-    {45.569, 52.8002, 148.5},     {30.5835, 60.3944, 148.5},
-    {30.5835, 60.3944, 220.5},    {45.569, 52.8002, 220.5},
-    {34.3926, 60.6757, -504},     {18.2134, 65.2005, -504},
-    {18.2134, 65.2005, -432},     {34.3926, 60.6757, -432},
-    {34.3926, 60.6757, -431.5},   {18.2134, 65.2005, -431.5},
-    {18.2134, 65.2005, -359.5},   {34.3926, 60.6757, -359.5},
-    {34.3926, 60.6757, 221},      {18.2134, 65.2005, 221},
-    {18.2134, 65.2005, 293},      {34.3926, 60.6757, 293},
-    {34.3926, 60.6757, 293.5},    {18.2134, 65.2005, 293.5},
-    {18.2134, 65.2005, 365.5},    {34.3926, 60.6757, 365.5},
-    {34.3926, 60.6757, 366},      {18.2134, 65.2005, 366},
-    {18.2134, 65.2005, 438},      {34.3926, 60.6757, 438},
-    {34.3926, 60.6757, 438.5},    {18.2134, 65.2005, 438.5},
-    {18.2134, 65.2005, 510.5},    {34.3926, 60.6757, 510.5},
-    {34.3926, 60.6757, -359},     {18.2134, 65.2005, -359},
-    {18.2134, 65.2005, -287},     {34.3926, 60.6757, -287},
-    {34.3926, 60.6757, -286.5},   {18.2134, 65.2005, -286.5},
-    {18.2134, 65.2005, -214.5},   {34.3926, 60.6757, -214.5},
-    {34.3926, 60.6757, -214},     {18.2134, 65.2005, -214},
-    {18.2134, 65.2005, -142},     {34.3926, 60.6757, -142},
-    {34.3926, 60.6757, -141.5},   {18.2134, 65.2005, -141.5},
-    {18.2134, 65.2005, -69.5},    {34.3926, 60.6757, -69.5},
-    {34.3926, 60.6757, -69},      {18.2134, 65.2005, -69},
-    {18.2134, 65.2005, 3},        {34.3926, 60.6757, 3},
-    {34.3926, 60.6757, 3.5},      {18.2134, 65.2005, 3.5},
-    {18.2134, 65.2005, 75.5},     {34.3926, 60.6757, 75.5},
-    {34.3926, 60.6757, 76},       {18.2134, 65.2005, 76},
-    {18.2134, 65.2005, 148},      {34.3926, 60.6757, 148},
-    {34.3926, 60.6757, 148.5},    {18.2134, 65.2005, 148.5},
-    {18.2134, 65.2005, 220.5},    {34.3926, 60.6757, 220.5},
-    {21.8946, 66.2195, -504},     {5.1435, 67.5009, -504},
-    {5.1435, 67.5009, -432},      {21.8946, 66.2195, -432},
-    {21.8946, 66.2195, -431.5},   {5.1435, 67.5009, -431.5},
-    {5.1435, 67.5009, -359.5},    {21.8946, 66.2195, -359.5},
-    {21.8946, 66.2195, 221},      {5.1435, 67.5009, 221},
-    {5.1435, 67.5009, 293},       {21.8946, 66.2195, 293},
-    {21.8946, 66.2195, 293.5},    {5.1435, 67.5009, 293.5},
-    {5.1435, 67.5009, 365.5},     {21.8946, 66.2195, 365.5},
-    {21.8946, 66.2195, 366},      {5.1435, 67.5009, 366},
-    {5.1435, 67.5009, 438},       {21.8946, 66.2195, 438},
-    {21.8946, 66.2195, 438.5},    {5.1435, 67.5009, 438.5},
-    {5.1435, 67.5009, 510.5},     {21.8946, 66.2195, 510.5},
-    {21.8946, 66.2195, -359},     {5.1435, 67.5009, -359},
-    {5.1435, 67.5009, -287},      {21.8946, 66.2195, -287},
-    {21.8946, 66.2195, -286.5},   {5.1435, 67.5009, -286.5},
-    {5.1435, 67.5009, -214.5},    {21.8946, 66.2195, -214.5},
-    {21.8946, 66.2195, -214},     {5.1435, 67.5009, -214},
-    {5.1435, 67.5009, -142},      {21.8946, 66.2195, -142},
-    {21.8946, 66.2195, -141.5},   {5.1435, 67.5009, -141.5},
-    {5.1435, 67.5009, -69.5},     {21.8946, 66.2195, -69.5},
-    {21.8946, 66.2195, -69},      {5.1435, 67.5009, -69},
-    {5.1435, 67.5009, 3},         {21.8946, 66.2195, 3},
-    {21.8946, 66.2195, 3.5},      {5.1435, 67.5009, 3.5},
-    {5.1435, 67.5009, 75.5},      {21.8946, 66.2195, 75.5},
-    {21.8946, 66.2195, 76},       {5.1435, 67.5009, 76},
-    {5.1435, 67.5009, 148},       {21.8946, 66.2195, 148},
-    {21.8946, 66.2195, 148.5},    {5.1435, 67.5009, 148.5},
-    {5.1435, 67.5009, 220.5},     {21.8946, 66.2195, 220.5},
-    {8.55507, 69.2185, -504},     {-8.12411, 67.2074, -504},
-    {-8.12411, 67.2074, -432},    {8.55507, 69.2185, -432},
-    {8.55507, 69.2185, -431.5},   {-8.12411, 67.2074, -431.5},
-    {-8.12411, 67.2074, -359.5},  {8.55507, 69.2185, -359.5},
-    {8.55507, 69.2185, 221},      {-8.12411, 67.2074, 221},
-    {-8.12411, 67.2074, 293},     {8.55507, 69.2185, 293},
-    {8.55507, 69.2185, 293.5},    {-8.12411, 67.2074, 293.5},
-    {-8.12411, 67.2074, 365.5},   {8.55507, 69.2185, 365.5},
-    {8.55507, 69.2185, 366},      {-8.12411, 67.2074, 366},
-    {-8.12411, 67.2074, 438},     {8.55507, 69.2185, 438},
-    {8.55507, 69.2185, 438.5},    {-8.12411, 67.2074, 438.5},
-    {-8.12411, 67.2074, 510.5},   {8.55507, 69.2185, 510.5},
-    {8.55507, 69.2185, -359},     {-8.12411, 67.2074, -359},
-    {-8.12411, 67.2074, -287},    {8.55507, 69.2185, -287},
-    {8.55507, 69.2185, -286.5},   {-8.12411, 67.2074, -286.5},
-    {-8.12411, 67.2074, -214.5},  {8.55507, 69.2185, -214.5},
-    {8.55507, 69.2185, -214},     {-8.12411, 67.2074, -214},
-    {-8.12411, 67.2074, -142},    {8.55507, 69.2185, -142},
-    {8.55507, 69.2185, -141.5},   {-8.12411, 67.2074, -141.5},
-    {-8.12411, 67.2074, -69.5},   {8.55507, 69.2185, -69.5},
-    {8.55507, 69.2185, -69},      {-8.12411, 67.2074, -69},
-    {-8.12411, 67.2074, 3},       {8.55507, 69.2185, 3},
-    {8.55507, 69.2185, 3.5},      {-8.12411, 67.2074, 3.5},
-    {-8.12411, 67.2074, 75.5},    {8.55507, 69.2185, 75.5},
-    {8.55507, 69.2185, 76},       {-8.12411, 67.2074, 76},
-    {-8.12411, 67.2074, 148},     {8.55507, 69.2185, 148},
-    {8.55507, 69.2185, 148.5},    {-8.12411, 67.2074, 148.5},
-    {-8.12411, 67.2074, 220.5},   {8.55507, 69.2185, 220.5},
-    {-5.11317, 69.5575, -504},    {-21.0795, 64.3311, -504},
-    {-21.0795, 64.3311, -432},    {-5.11317, 69.5575, -432},
-    {-5.11317, 69.5575, -431.5},  {-21.0795, 64.3311, -431.5},
-    {-21.0795, 64.3311, -359.5},  {-5.11317, 69.5575, -359.5},
-    {-5.11317, 69.5575, 221},     {-21.0795, 64.3311, 221},
-    {-21.0795, 64.3311, 293},     {-5.11317, 69.5575, 293},
-    {-5.11317, 69.5575, 293.5},   {-21.0795, 64.3311, 293.5},
-    {-21.0795, 64.3311, 365.5},   {-5.11317, 69.5575, 365.5},
-    {-5.11317, 69.5575, 366},     {-21.0795, 64.3311, 366},
-    {-21.0795, 64.3311, 438},     {-5.11317, 69.5575, 438},
-    {-5.11317, 69.5575, 438.5},   {-21.0795, 64.3311, 438.5},
-    {-21.0795, 64.3311, 510.5},   {-5.11317, 69.5575, 510.5},
-    {-5.11317, 69.5575, -359},    {-21.0795, 64.3311, -359},
-    {-21.0795, 64.3311, -287},    {-5.11317, 69.5575, -287},
-    {-5.11317, 69.5575, -286.5},  {-21.0795, 64.3311, -286.5},
-    {-21.0795, 64.3311, -214.5},  {-5.11317, 69.5575, -214.5},
-    {-5.11317, 69.5575, -214},    {-21.0795, 64.3311, -214},
-    {-21.0795, 64.3311, -142},    {-5.11317, 69.5575, -142},
-    {-5.11317, 69.5575, -141.5},  {-21.0795, 64.3311, -141.5},
-    {-21.0795, 64.3311, -69.5},   {-5.11317, 69.5575, -69.5},
-    {-5.11317, 69.5575, -69},     {-21.0795, 64.3311, -69},
-    {-21.0795, 64.3311, 3},       {-5.11317, 69.5575, 3},
-    {-5.11317, 69.5575, 3.5},     {-21.0795, 64.3311, 3.5},
-    {-21.0795, 64.3311, 75.5},    {-5.11317, 69.5575, 75.5},
-    {-5.11317, 69.5575, 76},      {-21.0795, 64.3311, 76},
-    {-21.0795, 64.3311, 148},     {-5.11317, 69.5575, 148},
-    {-5.11317, 69.5575, 148.5},   {-21.0795, 64.3311, 148.5},
-    {-21.0795, 64.3311, 220.5},   {-5.11317, 69.5575, 220.5}};
+extern std::vector<std::array<scalar, 3u>> odd_pixel_barrel;
 
 using rectangle = std::vector<std::array<scalar, 3u>>;
 
-inline static std::vector<rectangle> generate_barrel_modules() {
-    std::vector<rectangle> modules;
-    size_t number_of_modules = data::odd_pixel_barrel.size() / 4u;
-    modules.reserve(number_of_modules);
-    for (size_t im = 0; im < number_of_modules; ++im) {
-        modules.push_back({data::odd_pixel_barrel[4 * im],
-                           data::odd_pixel_barrel[4 * im + 1],
-                           data::odd_pixel_barrel[4 * im + 2],
-                           data::odd_pixel_barrel[4 * im + 3]});
-    }
-    return modules;
-}
+std::vector<rectangle> generate_barrel_modules();
 
 }  // namespace data
