+++ tests/common/detector.hpp
