+++ tests/core/CMakeLists.txt
