+++ web/CMakeLists.txt
@@ -2,5 +2,11 @@ message(STATUS "Building 'actsvg::web' component")
 
 # Set up the library.
 actsvg_add_library( actsvg_web web
-   "include/actsvg/web/web_viwer.hpp")
-target_link_libraries( actsvg_web INTERFACE actsvg::core)
+   "include/actsvg/web/web_builder.hpp"
+   "include/actsvg/web/webpage_text.hpp"
+
+   "src/web/web_builder.cpp"
+   "src/web/webpage_text.cpp"
+)
+
+target_link_libraries( actsvg_web PUBLIC actsvg::core)
