+++ web/include/actsvg/web/web_builder.hpp
@@ -22,39 +22,7 @@ namespace actsvg::web {
 namespace compare {
 // Function to compare two strings alphanumerically.
 bool alphanumeric(const actsvg::svg::object& svg1,
-                  const actsvg::svg::object& svg2) {
-    const std::string& str1 = svg1._id;
-    const std::string& str2 = svg2._id;
-    // Iterate through both strings simultaneously
-    size_t i = 0, j = 0;
-    while (i < str1.length() && j < str2.length()) {
-        if (std::isdigit(str1[i]) && std::isdigit(str2[j])) {
-            // If both characters are digits, compare them as integers
-            int num1 = 0, num2 = 0;
-            while (i < str1.length() && std::isdigit(str1[i])) {
-                num1 = num1 * 10 + (str1[i] - '0');
-                i++;
-            }
-            while (j < str2.length() && std::isdigit(str2[j])) {
-                num2 = num2 * 10 + (str2[j] - '0');
-                j++;
-            }
-            if (num1 != num2) {
-                return num1 < num2;
-            }
-        } else {
-            // Compare characters as alphabetic
-            if (str1[i] != str2[j]) {
-                return str1[i] < str2[j];
-            }
-            i++;
-            j++;
-        }
-    }
-
-    // Handle the case where one string is a prefix of the other
-    return i == str1.length() && j != str2.length();
-}
+                  const actsvg::svg::object& svg2);
 }  // namespace compare
 
 /// @brief Class for generating a web page to view and merge svgs.
@@ -95,13 +63,7 @@ class web_builder {
     }
 
     private:
-    void create_template(const std::filesystem::path& output_directory) {
-        create_directory_structure(output_directory);
-        write_file(output_directory / "index.html", index_text);
-        write_file(output_directory / "script.js", script_text);
-        write_file(output_directory / "styles.css", css_text);
-        write_file(output_directory / "rebuild.py", rebuild_text);
-    }
+    void create_template(const std::filesystem::path& output_directory);
 
     /// @brief Writes the svgs to the svgs folder in the output directory and
     /// generates the config.json file.
@@ -162,15 +124,7 @@ class web_builder {
 
     /// @brief Creates the folder structure for the webpage.
     void create_directory_structure(
-        const std::filesystem::path& directory_path) {
-        if (!std::filesystem::exists(directory_path)) {
-            std::filesystem::create_directory(directory_path);
-        }
-        const auto svg_directory = directory_path / "svgs";
-        if (!std::filesystem::exists(svg_directory)) {
-            std::filesystem::create_directory(svg_directory);
-        }
-    }
+        const std::filesystem::path& directory_path);
 };
 
 }  // namespace actsvg::web
